Allowed tabs to be marked as loaded
Plan to freeze loaded and inactive tabs. Decision of when a tab is first loaded is up to developer. Once a tab is selected it is considered loaded. A loaded tab can't be marked as unloaded. It's just about whether allowed to freeze - if a tab's been seen by user then can definitely freeze