Prevented freezing swipeable tabs within +/- 1
On android, can swipe tabs to left or right. Safest not to freeze these so don't get a flicker of old content when swiping either way. Mind you, similar scenario on ios with swipe back to frozen scene. But solve that by listening to peek and unfreezing. Don't want to do that yet because that would need a render when swiping. For now will just not freeze them