Ensured tab items removed cleanly
In Twitter example, tested swapping tab bar item children by showing either Home or Notifications in the home tab (on a timer)
(!change ? <Home />: <Notifications />)
But when the Notifications was removed got an error. When removing the second tab it updated the ViewPager which triggered getItemPosition - this threw an error because the first tab had no content. Guess the content is remove from both tabs first and then the tabs themselves are removed.
So added defensive check that the tab bar item is only changed if there is content. There always will be content (except this scenario)