Ensured anchored fab doesn't overlap buttons
On new architecture a fab anchored to bottom app bar and aligned to the end overlapped the buttons (didn't overlap on material 3 but think that's because end is the default alignment). I'm guessing this is because the order of events have changed but not sure. Requesting another layout when attached to window fixed it