Colored toolbar using searchbar barTintColor prop
The SearchBar material component doesn't have a way to set background color. Setting it on the view didn't work because it ignored the material rounded corner shape (it colored whole rectangle). So used the existing drawable to get the shape.
Made the barTintColor prop on SearchBar component a function (or string) so it works for both the toolbar and the results bar