Raised expanded and collapsed search bar events
This hides and shows the results content. Couldn't add and remove children like doing before because there's a 'bug' with the material SearchView. It overrides addView and puts them inside the container - but it doesn't override removeView so the results contents get added more than once.
Made zindex negative when closed so the results don't receive the press event