Ensured animated modal shows stack while sliding
When showing a modal with animationType slide the stack didn't appear until the slide animation completed. Tried it on medley sample and it worked, tracked the break down to dispatching the navigate async. This meant the modal animated played before the stack had time to render. So made the navigate sync - but only for the very first navigate. It's safe to dispatch subsequent navigations async and this is needed for the back button images - need to find the navigation bar so that has to happen after all views have loaded.