Ensured scenes view controllers are recycled
Tried the following in medley sample
1. Navigate to South
2. Open Modal
3. Close Modal
4. Navigate to West
5. Open Modal
6. Close Modal
Then the navigation bar in the West didn't show correctly and the scene contents weren't visible. Tried not recycling the Scene (using https://github.com/facebook/react-native/pull/35378/files) and the problem went away. So realised the scene recycling was wrong. Made it match the other UI Controller recycling (like Navigation and TabBar) and it worked!!