Fired onQuery for ios on fabric
Found problem that search button raised should end editing which set active to false which then cleared the text. But the search is still active, so changed to set inactive based on the cancel button click.
Found this on fabric only because non-fabric has but that event lag isn't 0 so wasn't being inactivated like on fabric. Need to move to didSetProps so event lag prop and active prop are in step