Copied over fabric changes
Set inactive when cancel clicked so that search enter key press doesn't fire it and clear the text. Also ensured that event lag is updated when active check runs - this error was brought in when updating to fabric. The event lag stuff is not done using setNativeProps anymore so no guarantee the event lag is updated before the active check - so do this change for all event lag checks on old architecture