Decided which hide based on back title
Don't want to rely on navigation bar not being there because the status bar only works if the navigation bar is there (what if want not navigation bar, swipe back and status bar). Instead using the back title to decide which type of hide. Hidden with no back title is the default current behaviour. Hidden with back title keeps the swipe