Applied different hidden in navigation bar too
Could, in theory, keep hidden true and toggle backTitle so can enable and disable the swipe. So need to apply different hidden settings inside the NavigaitonBar too. Changed property from hidden to isHidden because hidden is an iOS property already and it kept being set to true when toggling backTitle even if hidden prop stayed false