Applied same changes to bottom sheet
This fixed the event lag being out but then that set the detent to an undefined value when user scrolling. Should ignore the scrolling detent, but the event count state update triggered update - which still had the old detent (it ignored undefined detent). So moved the event count update into the same check so it doesn't run if detent is undefined