Prevented losing landscape when going back
Navigate from A --> B --> C with both B and C in landscape, then going back to B goes to portrait (unspecified). The getAnimation was returning null because the crumb was less than the current crumbs so it went back to default orientation.
Changed to use the navigationEvent from state because that doesn't change when navigating - that's fixed for current scene. It's not guaranteed to be there, for example, fluent navigation, so used the navigationEvent from props if the one from state isn't there