Suppported landscape on ios (< 16)
Think the simulator behaves oddly so can't be 100% it's correct. There's interface orientation and device orientation. When rotating the simulator that changes both of these. When programmatically changing the orientation it also seems to change both but without rotating the simulator. So it's hard to know what position the 'real' device is in. But assuming that the device orientation is reliable indicator of its position. So when changing to landscape programmatically the device is actually rotated (even though simulator isn't). Basically can't replicate orientation landscape but device portrait in simulator