Rewrote urls and cached the result for get href
Tried putting the rewrite logic into get href of history manager but this changes the parameters and so it can't be used For example, in react native web sample there's manual calls to get href and these wouldn't work if need to pass in more params.
Parsing the url to get the state and data back in get href would work but not allowed to call the slow parse for every link. Can only call this expensive op when navigating - not when building links.
There's always a call to get link before a call to get href so can cache the get href call while have access to the state and data. Then make the rewritten cache available to the history manager