Rewrote the urls in state context on navigation
Need to rewrite the navigated url so that can get href in history manager. But also best to rewrite the crumbs to because might want to get their hrefs and won't have called get link on them first, for example, if navigating a deep link.
Deleted the cache on navigation too. Don't want cache to just grow forever. Plus it should be safe to delete the cache because get link should always be called before get href - except for the urls in state context so adding them all to cache after deletion