Used handler instead of fluent
Fluent is going to call rewrite so don't want infinite loop. But also it's better. Was only used fluent when needed to rewrite from the links. Now it's all in state navigator no need for fluent call.
Also means that the invalid test doesn't throw anymore so it writes the original url if the rewrite is missing a route parameter