Passed crumbs to rewrite
Allows full rewrite of stack just based on the current url. Don't want to pass full state context because old state, for example, is different depending on how the link is clicked - crucial that can do the rewrite from anywhere, not dependent on journey taken. All info contained in the url to rewrite. Also allows user to change the stack although typically should return the crumbs passed in - or no crumbs - the same outcome if using clean urls because history manager will strip the crumbs anyway