Rewrote the crumbs
Navigate from A --> B, rewrite A to C and B to D. Then the href should be C --> D not A --> D. So rewrote the whole href including the crumbs. Only change the crumbless Url but kept the state, data and url on the crumbs the same - same as how state and data aren't change on state context. The rewrite only impacts the href and nothing else.