Put rewritten state and data in crumbs
Truncate crumb trail is called on the rewritten state and data - it has to be. So the same should happen for the crumbs. Gave them the rewritten state and data. Didn't give them the rewritten url though - it's a bit harder and think it's right to give them the original url