Removed crumb rewrites
Couldn't see what is was for. If navigate from A --> B --> C with B rewritten to D without trail then should that go to D --> C? Wasn't doing that but could make a case for it. Better to rewrite just the current state and data - clearer to understand and less calls to rewrite too. Wasn't happy rewriting every crumb because could lead to a lot of rewrites - for no gain