Removed parallel route test
Can't really think of a valid use-case. Why not store the other tab in the query string. Don't see a reason for the shareable link to only include one tab.
Also messes with the active link prop. Need to store an item that indicates the last selected tab so selected tab but not last selected isn't marked as active. Can revisit if there's a valid usecase