Renamed for clarity
It's not route interception because the Navigation router doesn't care if it's route or not. It's better to call it rewrite but intercept is the common use-case (from Next examples). The navigation isn't being intercepted. The Navigation is always to the Modal. The navigation is being rewritten when made into an href