Truncated crumbs before rewriting
The crumbs didn't work because it depended on the direction taken. For example navigate A --> A or refresh A would pass different crumbs even though the url is the same (if track crumb trail is false). In the first it would pass crumbs [A] but the second would pass []. So truncated crumb trail so that the crumbs are always the same - in this example, would be [] in both cases