Removed crumbs when rewriting
Crumbs just don't work because they're dependent on the route taken. For example, A --> A and refresh A has different crumbs (track crumb trail false). In the first the crumbs is [A], in the second it's empty. There's no advantage to passing crumbs but thought it worked. Don't need them because want clean urls when rewriting