Brought crumbs back when rewriting
As long as truncate them and clear them if not tracking then they'll be the same no matter what direction is taken.
The example of A --> A and refresh A having different crumbs - when track crumb trail is false - won't happen anymore. They'll both have empty crumbs
If track crumb trail is true then they will have different crumbs - that's ok because the unrewritten urls are different.
The key is that if the url is the same then the rewritten url must be the same too.