Got zoom sample working with original shared el
Original shared element means not the new material container transform shared element. The ChangeTransform transition caused the old shared element to disappear when remapping - so click red on grid, click brown on details, click back and red would be invisible. Had to override endViewTransition in SharedElementView and force it back to visible