Faded the enter/exit scenes when shared elements
Custom animations don't run when there are any transitions (for example, shared el). Need to use transitions instead. But then there are 2 problems

When going back the disappearing scene content disappears. The fragment is destroyed immediately without waiting for the transition to complete. So the scene is popped and removed on the React side which removes it on native. Delayed the popped event call until the transition completes. The Enter transition runs when popping too (when no return transition is specified).

Even with this in place any image on a fragment transitioning away (into stack or off the stack) disappears. Fixed this by calling startViewTransition on the scenes. This prevents them being removed from their parent NavigationStackView and so images don't detach and stay visible.