Ensured rested after shared element transition
Tried to call onRest from the fragment but the scene parent is some overlay so can't get to the NavigaitonStackView from there. Moved the transition listeners into the stack so can call onRest. Had to put the onRest in both enter and exit transitions so that it's called with the correct crumb - enter when forward and exit when back