Ensured style changes happen on first render
With duration 0 navigate from A → B and swipe back on Safari. Letting go of the swipe would show a flash of scene B before A. The `marker` meant that B wouldn’t leave on the first render.
Added the `marker` in 1accf8e to ensure scenes come to rest at the same time. If the mounted scene rests while another is still animating then the mounted scene would keep rendering. But in 8e3197d changed to wait for all scenes to come to rest before rendering the mounted scene (to handle different durations). So should’ve removed the `marker` as part of that anyway.