Dynamically changing icon animation didn't work
Set navigation icon to home.png then open and closed the search, then removed the navigation icon and opened search again - the home icon was still shown when animating between toolbar and results. So simplified code by setting the animation just when attached to window. It doesn't fix the bug but no point updating each animation when it doesn't work