Removed redundant pause listener
It threw an error when pressing very fast going forward with long duration transition - because the parent was an Overlay and not the NavigationStackView. Also it's not needed because if it's paused it means another transition has taken over so can let that one call rest. Only need the latest enter transition to call rest - previous ones in flight aren't relevant