Ensured onRest is always called
If there's an exit transition but no enter one then rest wasn't called. Rest was tied either to animation end or transition end. But exit transition only means no animations and no enter trans.
So checked for this and passed 0 animation so that it goes down the instant onRest call