Prevented live reload giving blank screen
In Twitter sample, navigated A --> B then back to A. Then edited NavigationStack underlay color and navigated to B. This would give empty screen.
React Native reruns the App when editing the App.js which caused an empty navigation. Back the crumb was already on 1 so it tried to go back. And that removed the view and gave a blank screen. Not totally sure of the reason - but basically a few navigations happen when editing App.js then navigating.
Only removing the parent if not the stack solved it. This also still works for when fast navigating - reason this code was introduced. Since that only failed because tried to set parent when it was already the overlay because still animating