Allowed toggling bottom and primary tabs
The tabs would disappear, for example, when toggling from bottom to top tabs. The tab view would go blank because it wasn't rebound to tab bar. The tab view was attached to the window but it wasn't bottom so it didn't rebind.

Removed all the checks around rebinding so that the tab view and tab bar always rebind on attached to window. Had to stop clearing the menu in TabNavigationView on rebind otherwise the selection got lost. Instead updated the menu items if already there.

Had to change the way the TabLayout worked out whether it was the one in the NavigationBar or not. Was hiding it and then showing it, but this ended up showing it when dropped the bottomTabs check. Changed to check if it's the one the in the navigation bar. Copied the check from the TabBarPagerView.

The selection was lost when toggling between primary and not. The mostRecentEventCount was remembered but this is state in the component, but the nativeEventCount was lost because this is held in native. Different native components. So set native from mostRecent if mostRecent is bigger.

Gave the components keys so that they're not destroyed. Noticed the selection was lost. But this probably unnecessary now doing the event count.