Prevented infinite loop if change tabs as toggling
It's possible to get infinite loop if change tabs fast enough from javascript (controlled tab prop). The native event doesn't help here because the change is triggered from javascript, not native - change to tab 2 in javascript, before the change comes back to native change to tab 1 in javascript, then the change from tab 2 comes back and event goes to javascript, and change from tab 1 comes back and event goes to javascript etc.

Found this scenario when changing tab as toggle - not the same because the change here is from native. But ignore the native event count so it acts like the same scenario. Have to ignore the native event so that the selected tab is remembered when toggling (the native event count is reset when toggling).

Added jsUpdate check so don't send the event back to js if it came from js. So if the change happens in onAfterUpdateTransaction then don't tell javascript about it