Supported toggling primary
Toggle the primary prop in useEffect on a setTimeout. Select the notifications tab then when it toggles to non-primary the notifications tab doesn't load - it's a black hole where the scene should be. It worked if disabled view recycling for the SceneComponentView (hard-coded in RCTComponentViewRegistry).

Weirdly deleting the contents of ensureViewController fixed the toggling. But it broke [the reason it was introduced](https://github.com/grahammendick/navigation/commit/f5c0c62b8113f6691549a10323ce08ddf63fef37)
Tried the following in medley sample
1. Navigate to South
2. Open Modal
3. Close Modal
4. Navigate to West
5. Open Modal
6. Close Modal
Then the navigation bar in the West didn't show correctly and the scene contents weren't visible.

Moving the ensureViewController code into the prepareForRecycle fixed toggle and medley. Guess it's a timing thing. the ensure was called much later than the prepare - no idea really, it's mostly just trial and error