Ensured freezable set to true if not selected
Need that non-selected tabs are freezable after first rendering. User might call onLoad for lazy loading a tab and it should freeze - only works if freezable set to true. Used to start as true and useEffect in tab bar item would freeze - but freezable has to start as false because it has to unfreeze when tab count changes