Prevented infinite loop on fabric
On Android select tab 2 and press back then it would go into an infinite loop. Pretty sure there's a fabric bug with suspense.  Also saw infinite loop on ios when adding a tab - fixed this by unfreezing when number of tabs change. Think fabric gets confused when a state update comes in for a suspended component?! not sure.

Anyway, fixed this by delaying the freeze for a render