Supported remove tab and controlled change on ios
Had 3 tabs and in useEffect removed the 3rd and selected the 1st. Before the useEffect ran, manually went to the 2nd and did some navigations (3 or more typically but not sure it mattered). When the useEffect ran the 1st tab was selected correctly but when went to the 2nd tab and pressed back through the stack it had too few screens!!! could go back 2 and ios would think it could go back again - would show back arrow - but the stack didn't have enough screens so got stuck. Seems like one the view controllers was removed. So there weren't enough to go back through.

Disabled the recycling and it worked. So reverted change made to support toggling primary where removed the in the prepareForRecycle and put it back into ensureViewController. And it worked!

Then tested the reason moved it in first place (below) and it was ok. I'm guessing the real problem was the suspending but didn't realise originally. Now delayed the suspending the toggling primary problem isn't there anymore.

> Toggle the primary prop in useEffect on a setTimeout. Select the notifications tab then when it toggles to non-primary the notifications tab doesn't load - it's a black hole where the scene should be.