Ensured controlling tab when removing tab
Start with 3 tabs with 1st tab selected. Then in useEffect remove the 2nd tab and set the tab to new 2nd programmatically.
This threw error sometimes on primary tabs and flicked back to select 1st tab on non-primary. Turned off freeze to test and it worked fine.
Prevented sending the update back to TabBar - it's not necessary and think React does this anyway for it's controlled props. Found this out when investigating https://github.com/grahammendick/navigation/issues/719. It's better not to send updates if can help it but fabric really hates these updates to suspended components.
Still send the pressed event - seems fine and do want the pressed to fire because might be hiding badge for example