Ensured bottom sheet supports reload
Tapping R (to reload) didn't dismiss the bottom sheet. Meant that it the old sheet was still around 'unattached' to react. Don't know why React Native doesn't remove from the view and invalidate?! It does invalidate the manager. So fired a notification from there and listened to it in the view and dismissed the sheet.
Checked React Native's approach for Modal. The manager retains pointers to the views so it can dismiss. This sounds bad. Don't think Modals are ever deallocated?! Don't know why they didn't used notifications.