Removed prop else sheet stays across navigation
If have largest undimmed the scene behind remains active. So can navigate and the sheet stays open?! but the sheet is controlled by a scene that's a crumb so it's suspended. What should happen here?! Removing for first version and can think about it more. Without undimming the sheet is modal so scene behind isn't active