Started the dialog bottom sheet on android
Added a modal prop and rendered dialog fragment if true. Need to copy the react native modal because it's a full screen dialog - need to resize it based on the full screen. It's not like the non-modal sheet that is sized relative to its container. So gonna need shadow node stuff