Ensured view takes up the whole dialog
If the dialog is shown from within a tab then the dialog took up whole scene but the view didn't fill it - was a tab sized gap at the bottom.
Thought needed to do the shadow node stuff from React Native Modal but didn't make any difference. Updating the node size was enough to fill it (copied the host view and resize call from the Modal)