Fixed yoga/style/Style.h not found with install
React Native [recommended install_modules_dependencies](https://github.com/reactwg/react-native-new-architecture/discussions/115) but know why. Turns out it fixes the Yoga  style not found - had to clear out the other dependencies in podspec for it to work. The React Native doc struck through the recommendation but does still suggest it so confusing - but saw that svg and screens both have it and they work on react native 73 so that encouraged me to keep going with it