Removed dependency on namespace restricted code
ActionMenuItemView is restricted access so was getting red lines in android studio. Found another way to get the test id from the view - gave the menuitem a specific id and this gets copied through to the created view. Used the bar button id so will get duplciate ids if action views used - think this is ok but will test