Displayed sheet undimmed when modal is false
When undimmed can tap the background so dismissed the sheet when navigating back or forward. Important that the sheet is only open for the current scene. For example, the scene is frozen when in the stack so can't have the sheet open.
React Native calls invalidate when navigating back so dismissed the sheet there. For forward, useUnloaded