Ensured sheet opens when go back from sheet
Open sheet, navigate forward, open sheet, navigate back - this should open the first sheet, but it didn't. The present fired before the dismiss - not sure why that mattered because they're different view controllers - but it did. Dispatched immediately if on main thread then the order is correct. Only won't be on main thread if double tap R to relaod the whole app