Ensured sheet dismissed when navigated away
Can't rely on prepareForRecycle because view not recycled if pool size reached. So dismissed when moved out of window. Also invalidate the displayLink otherwise the view isn't deallocated (if not recycled).
The displayLink is hooked up again when added back to window - because the sheet is recreated (js sets dismissed and returns null when navigated away from)