Matched dismissal to fabric
Couldn't get the dismiss to be animated on fabric - the timing just never worked out. The content was gone before the animation so it went blank as animating away. Even tried taking snapshots - this fixed the blank, but didn't work when opening 2 sheets in a row on different scenes. When navigating back the second sheet didn't animate. Think the first sheet reappearing interrupted it somehow.
So prefer to keep old and new arch the same. Also meant could simplify the double tap 2 to reload. Don't need notification center to communicate from manager to view - dismissing when removed from window handles this case, too