Hid/showed sheet(s) when navigated away/back
This took tons of head-scratching. The problem is that non-modal sheets don't disappear when navigating. Doesn't make sense to leave them open though because they're part of another scene (also that old scene is frozen anyway).
But very trick to dismiss the sheets because there's lots of timing issues. For example, Open sheet in scene A then navigate to B with sheet of its own. Can't dismiss all sheets because then the sheet in B won't open. Can't dimiss just sheet in A because that dismisses sheet in B too. Super messy. General idea is to dismissAll when presenting and only call super present when the dismissAll completes. Also call dismissAll whenever navigating so always closes. Kept the sheet in memory when navigating back by holding reference to view controller. Otherwise the view is unloaded and the sheet goes invisible as it navigates away.

Also reopening sheets when navigating back is complicated. Tracked dismissed separate to detent - calling dismissAll doesn't call didDismiss so can send dismiss sepatately. Then when navigate back can see if dismissed but detent not hidden and remove dismissed to reshow.

Also had to still render the sheet when dismissed because want to support multiple sheets at once. For example, open sheet 1 then sheet 2 (inside sheet 1 - have to be nested). Then close sheet 2 and navigate forward and back. Only want sheet 1 to reopen. If return null from component when dismissed then the state of the nested sheet is lost and it would reopen. So rendered with negative zindex so it keeps react state but isn't touchable. Tried using height 0 but didn't work!? maybe because using c++ state for height?!