Ensured sheet view is deallocated
Can't guarantee that React Native recycles the view (there's a pool size). Need to make sure it's garbage collected if it's recycled. Turned off recycling in RCTComponentViewRegistry and added dealloc call and breakpoint and it's the displayLink reference that prevented it hitting the breakpoint. So invalidate the link when dismissed and added back when presented