Prevented sheet contents disappearing
On scene A open a sheet then navigate to scene B with its own sheet then navigate back - the 2nd sheet with go blank as it animates away.
The dismiss view controllers called when reopening scene A sheet happens first then the dismiss from the didShowViewController happens second. That meant that the completion was instantly called and the disappearing view controller was deallocated and the scene popped - making the contents go blank.
Changed so that the first call to dismiss keeps hold of the disappearing view controller and the dismiss completes