Ensured reload (tap r) works with first sheet
When sheet on first scene and tap r on a scene with a sheet on then it reloaded funny. The sheet would open fully expanded (and nested sheet didn't appear). The sheet could even have been dismissed as long as tap r on scene with sheet open (and sheet opens on first scene load).
Immediately dismissing the sheet on recycl/dealloc fixed it. Guess the animation interfered with the reload - maybe overlapped?!