Returned null when dismissed by user/dev
If dismissed by user or dev then the selected detent is hidden. If it's dismissed by navigation then selected detent won't be hidden.
So safe to return null if user/dev dismissed. That's how it worked before this PR and it's how the react native modal works. Means that the screen state is reset if reopened which makes sense