Ensured detent preserved when switch modal flag
Start modal and change detent then change to non-modal and detent would be forgotten. The mostRecentEventCount comes from js so is remembered but the nativeEventCount comes from native so is forgotten (because native component changes). So ensured native event count is the max because should never be behind.
Also ensured dialog takes the detent on first load otherwise it still didn't remember