Ensured back works after fluent navigation
Fluently navigate from A to A -> B -> C then navigate back. This fell over because the reenter animation for B was null when using the fluent api. When fluently navigating to C the prevFragment of B wasn't in the fragmentManager because the commit is async. So stored the prevFragment as went along instead - no need to get from the FragmentManager for fluent