Passed enter and exit trans props to ios
Had to set up the native component spec otherwise react native doesn't pass the prop values. It doesn't matter on Android - react native passes them without worrying about the spec. In fact, I tried to set up the spec when doing animation api on android and couldn't get it working. But suspect that was because using an old version of react native because it's working now.
But another problem was that number values inside the complex object weren't passed properly. For example, set a duration in the first item then react native kept that duration value for the other items when it was undefined on them. Also passing undefined or even default values didn't work. So changed all the values to strings because react native handles strings correctly. Will convert them on the native side