Removed unnecessary transition props
If enter transition then the to values aren't needed (and vice versa). It's always going to the identity. So only send and stored minimal info for clarity.
On Android it's simplest to send both because build animations directly out of from and to values. But on ios don't build animations until transitioning, at which point know the direction so don't need to store the direction on the transition obj