Added pop exit/enter transitions and transforms
Had to insert the subview below so that the exiting scene appears on top. Also had to set the transform back to identity of the crumb scene so that the transforms are applied from scratch when it returns. Otherwise it remembers where it was.
Maybe don't have to apply the transforms for the crumb if it starts from where it left off - but simpler to apply, plus user might not want it to start from where it finished