Implemented animation duration
At the moment not doing different durations for separate animations within a single enter/exit transition. Can have different durations for the enter and exit but not within each one. So take the max of all the transitions in the array. Defaulted to 300 to match android.
Ran the completion when the longer of enter or exit transition completes