Backed out the swipe back PR #677
Used to be that when navigation bar was hidden the swipe back gesture was disabled. So put in #677 to hide navigation bar in a different way to keep the swipe back (turned this on when back title was non-empty). But now the NavigationStack implements UIGestureRecognizerDelegate, and returns YES from gestureRecognizerShouldBegin, the swipe back is automatically enabled even when the navigation bar is hidden. So backing out #677 (on new arch, will do old arch in later commit)