Implemented fluent animations
Navigation from A to A -> B -> C and need the animations to work as navigate back through the stack. Ensured applied the pop enter and pop exit to the previous scene controller as loop through. The previous defaults to the top view controller - for the non-fluent case - and otherwise refers to the previous one from the loop. This matches the android implementation with prevSceneFragment