Implemented replace animations
Navigate from A -> B to A -> C and need the right animations to play. So C should play the enter animation B should play the pop exit animation. Well, I think it should be ios by default plays the exit animation for C - so it looks like C is going back into the stack (like a crumb when a push happens). This isn't right because C is off the stack - so should play the pop exit because C isn't in the stack anymore. Also, no idea what the exit animation is because don't work that out until another scene D is pushed on top - the crumbStyle needs the next state to work out what the animation should be. So there's no way to play the exit animation - but sure it's wrong and should play the pop exit because it's being popped. So set the exit from the pop exit - know what the pop exit is because that doesn't rely on the next state (can be popped without a next state)!
Also set the transitions onto the current and prev controller - matching android behaviour. The prev is not the top because the top isn't in the stack. The prev is the 2nd in the current stack because that's the previous in the new stack (the React side works on the stack only)