Cancelled the other gesture if swipe back
This is simpler way to ensure get the default swipe back behaviour. Need to be able to diagonally swipe back like default - so need to fail the scroll view pan if it's the interactive pop. Not sure if cancelling all others instead of just scroll (like react native screens) is right?! but think react native screens only cancels scroll view pan because it needs pan to pop as well