Implemented basic custom swipe back gesture
Added gesture recognizer for the left edge and started this one instead of the default interactive pop gesture recognizer. Created a percent driven interactive transition and updated as swipe progresses.
Sometimes the swipe back doesn't work!? It just does nothing. I'm blaming the simulator for this because pretty sure the code is right!?