Ensured can always swipe back
Blamed the intermittent failure on the simulator but it wasn't that. Was testing in the twitter sample and that has nested NavigationStacks - both of them have gesture recognizers. If the top level one fires first, it cancels the actual one and so no swipe back happens. Need to support this nested stack - could be stack inside modal inside stack etc.
So decided which gesture recognize to fail (cancel) based on the hierarchy. So if stack A contains stack B then in the fail method of A return NO so that gesture recognizer B responds. And in fail method of B return YES so that gesture recognizer A doesn't respond anymore. Used parentViewController to find ancestors.
(thought about return NO from should begin but too hard to find the hierarchy in here - advantage of the fail method is that given both the stacks!)