Animated only if the scene on top is animating
On Android can have a mixture of the default platform animation and a custom one. For example, A -> B can custom animate A and have default animation for B. But can't do that on iOS - can't have half, either both scenes custom or both scenes default. If only A is custom then doesn't make sense to run custom because wouldn't be able to see it - B would immediately cover the screen.
So only run custom if the scene on top has custom animation - otherwise use the default animation. The scene on top is the to scene when pushing and the from (top) scene when popping.
This also works better when turning off animation (using empty string). By default iOS will play the default animation when popping even when the scene was pushed without animation. So want to keep that idea even if custom animations are there for the crumb scene. Checking the top scene for custom means it will use the default if none are provided - matching the default platform behavior