Checked peekability in did begin in old arch
If the previous scene isn't loaded yet then return false. This happens if fluently navigating and then try to go back before the background peek has run. Used to have to disable the interactive gesture recognizer but simpler to return false from should begin - because that's called when the peek is tried