Disabled custom animations on ios by default
For backward compatibililty can't have custom animations on by default. So added customAnimation prop to stack that defaults to true on Android and false on iOS. If it's true then sends undefined for enter/exitAnim/Trans. Unless it's empty string because still want to be able to turn off animations when using default animations