Ensured transitions work with latest fragments
Same change as already done for forward navigation now doing for replace. If navigate from A -> B to A -> C with crumbStyle transition and unmountStyle blank then the fragment when blank if cancel the gesture back. Can't get transitions working at all with predictive back but ok to release without transition support (as long as transitions work as they do now)